pyproject.toml
[tool.poetry]
name = "repl-nix-blobstorage-webpage"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]
readme = "README.md"
packages = [{include = "repl_nix_blobstorage_webpage"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
azure-devops = "^7.1.0b4"
config = "^0.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
